chore: reorganize successful system logic
