+++ packages/client/src/layers/backend/systems/createSuccessfulActionSystem.ts
@@ -1,11 +1,9 @@
-import { defineComponentSystem, defineRxSystem, EntityIndex, setComponent } from "@latticexyz/recs";
-import { ActionState } from "@latticexyz/std-client";
+import { defineRxSystem, EntityIndex, getComponentValueStrict, setComponent } from "@latticexyz/recs";
 import { BigNumber } from "ethers";
 import { BytesLike, defaultAbiCoder as abi } from "ethers/lib/utils";
-import { merge } from "rxjs";
 import { ActionStruct } from "../../../../../contracts/types/ethers-contracts/ActionSystem";
-import { ActionType, Move } from "../../../types";
-import { BackendLayer, TxType } from "../types";
+import { ActionType } from "../../../types";
+import { BackendLayer } from "../types";
 
 export function createSuccessfulActionSystem(layer: BackendLayer) {
   const {
@@ -20,38 +18,12 @@ export function createSuccessfulActionSystem(layer: BackendLayer) {
     components: {
       ExecutedExtinguishFire,
       ExecutedShots,
-      EncodedCommitment,
-      CommittedMove,
       ExecutedChangeSail,
       ExecutedLoad,
       ExecutedRepairCannons,
       ExecutedRepairSail,
-      LocalHealth,
     },
-    actions: { Action },
-    utils: { clearComponent, isMyShip, playSound },
-    godIndex,
   } = layer;
-
-  defineComponentSystem(world, Action, ({ value }) => {
-    const newAction = value[0];
-    if (!newAction) return;
-
-    const state = newAction.state as ActionState;
-    if (!newAction.metadata) return;
-    const { type, metadata } = newAction.metadata as { type: TxType; metadata: any };
-    if (type == TxType.Commit) {
-      const { moves, encoding } = metadata as { moves: Move[]; encoding: string };
-      setComponent(EncodedCommitment, godIndex, { value: encoding });
-
-      moves.map((move) => {
-        const shipEntity = world.entityToIndex.get(move.shipEntity);
-        const moveCardEntity = world.entityToIndex.get(move.moveCardEntity);
-        if (!shipEntity || !moveCardEntity) return;
-        setComponent(CommittedMove, shipEntity, { value: moveCardEntity });
-      });
-    }
-  });
 
   function parseLoadAction(action: BytesLike) {
     const [cannonEntity] = abi.decode(["uint256"], action);
@@ -70,13 +42,12 @@ export function createSuccessfulActionSystem(layer: BackendLayer) {
     return { cannonEntity, targets };
   }
 
-  defineRxSystem(world, merge(systemCallStreams["ds.system.Action"]), (systemCall) => {
+  defineRxSystem(world, systemCallStreams["ds.system.Action"], (systemCall) => {
     const { args, systemId, updates } = systemCall;
     const { actions: rawActions } = args as {
       actions: ActionStruct[];
     };
 
-    const healthUpdates = updates.filter((update) => update.component == Health);
     rawActions.forEach((action) => {
       const shipEntity = world.entityToIndex.get(bigNumToEntityID(action.shipEntity));
       if (!shipEntity) return;
@@ -91,12 +62,12 @@ export function createSuccessfulActionSystem(layer: BackendLayer) {
           if (!cannonEntity) return;
 
           const damage = targets.map((target) => {
-            const newHealth = healthUpdates.find((update) => update.entity == target)?.value?.value as
-              | number
-              | undefined;
-            return 2;
-            // const oldHealth = getComponentValueStrict(LocalHealth, target).value;
-            // return oldHealth - (newHealth || oldHealth);
+            const newHealth = updates.find((update) => update.entity == target && update.component == Health)?.value
+              ?.value as number | undefined;
+            const oldHealth = getComponentValueStrict(Health, target).value;
+
+            console.log("old health:", oldHealth, "new health:", newHealth);
+            return oldHealth - (newHealth || oldHealth);
           });
           console.log("targets:", targets, "damage:", damage);
           setComponent(ExecutedShots, cannonEntity, { targets, damage });
